Removes the need to run provision again. Removes warning about key ID.
It resolves the issue of running provision after a setup. The problem was caused
when installing java.
See: https://askubuntu.com/questions/769467/can-not-install-openjdk-9-jdk-because-it-tries-to-overwrite-file-aready-includ
Adds full fingerprint for _tinyprod_ source and as such it removes the warning
about having a short ID.
Puts the correct location of _profile_ in _source_